Added "Inline Images" these are small images which are directly embedded in the XML when saving a Document. When reading such a Document they are extracted into temporary files which are deleted if the corresponding PageItem is deleted or another image is loaded into the frame. These inline images are currently only used by the SVG Importer.
ScribusXMLDoc::WriteElem() adds now a small preview image to the xml, this image is used by ScPreview to render
the icons for the scrapbook.

git-svn-id: svn://scribus.net/trunk/Scribus@12795 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/collect4output.cpp
@@ -189,20 +189,23 @@ void CollectForOutput::processItem(PageItem *ite)
 	if (ite->asImageFrame())
 	{
 		/* hack for subsequent c4o "./" -> "/doc/full/path" */
-		QString ofName(ite->Pfile);
-		QFileInfo itf = QFileInfo(ofName);
-		if (!itf.exists())
+		if (!ite->isInlineImage)
 		{
-			ofName = QDir::convertSeparators(PrefsManager::instance()->documentDir() + "/" + ofName);
-			itf.setFile(ofName);
-		}
+			QString ofName(ite->Pfile);
+			QFileInfo itf = QFileInfo(ofName);
+			if (!itf.exists())
+			{
+				ofName = QDir::convertSeparators(PrefsManager::instance()->documentDir() + "/" + ofName);
+				itf.setFile(ofName);
+			}
 		// end of hack
-		if (itf.exists())
-		{
-			QString oldFile = ofName;
-			ite->Pfile = collectFile(oldFile, itf.fileName());
-			ScCore->fileWatcher->removeFile(oldFile);
-			ScCore->fileWatcher->addFile(ite->Pfile);
+			if (itf.exists())
+			{
+				QString oldFile = ofName;
+				ite->Pfile = collectFile(oldFile, itf.fileName());
+				ScCore->fileWatcher->removeFile(oldFile);
+				ScCore->fileWatcher->addFile(ite->Pfile);
+			}
 		}
 	}
 	if (ite->asTextFrame())
@@ -212,13 +215,16 @@ void CollectForOutput::processItem(PageItem *ite)
 			QFileInfo itf;
 			if (!ite->Pfile.isEmpty())
 			{
-				itf = QFileInfo(ite->Pfile);
-				if (itf.exists())
+				if (!ite->isInlineImage)
 				{
-					QString oldFile = ite->Pfile;
-					ite->Pfile = collectFile(oldFile, itf.fileName());
-					ScCore->fileWatcher->removeFile(oldFile);
-					ScCore->fileWatcher->addFile(ite->Pfile);
+					itf = QFileInfo(ite->Pfile);
+					if (itf.exists())
+					{
+						QString oldFile = ite->Pfile;
+						ite->Pfile = collectFile(oldFile, itf.fileName());
+						ScCore->fileWatcher->removeFile(oldFile);
+						ScCore->fileWatcher->addFile(ite->Pfile);
+					}
 				}
 			}
 			if (!ite->Pfile2.isEmpty())
